LUCENE-4635: fix overflow bug causing AIOOBE when index has many unique terms
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene_solr_3_6@1423711 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/PagedBytes.java
@@ -411,7 +411,7 @@ public final class PagedBytes {
 
     /** Returns the current byte position. */
     public long getPosition() {
-      return currentBlockIndex * blockSize + currentBlockUpto;
+      return ((long) currentBlockIndex * blockSize) + currentBlockUpto;
     }
   
     /** Seek to a position previously obtained from
@@ -516,7 +516,7 @@ public final class PagedBytes {
       if (currentBlock == null) {
         return 0;
       } else {
-        return blocks.size() * blockSize + upto;
+        return ((long) blocks.size() * blockSize) + upto;
       }
     }
   }
